feat(consulting): redesigned consulting page with animated terminal and Cal.com booking
  Complete overhaul of consulting page with personality-driven design. Replaced
  plain text with animated terminal showcasing real experience (Cosmos validator,
  database-first architecture, AI/LLM engineering). Integrated Cal.com modal
  booking system. Split into server/client components following Next.js App
  Router patterns with SSG (revalidate: false).
