+++ supabase/functions/llms-txt/index.ts
@@ -101,7 +101,7 @@ Deno.serve(async (req) => {
       }
 
       case 'item': {
-        // /[category]/[slug]/llms.txt - Individual item (existing function returns JSONB)
+        // /[category]/[slug]/llms.txt - Individual item (returns pre-formatted TEXT)
         if (!category || !slug) {
           return badRequestResponse(
             'Missing required parameters: category and slug',
@@ -112,14 +112,14 @@ Deno.serve(async (req) => {
           return badRequestResponse(`Invalid category: ${category}`, getCorsHeaders);
         }
 
-        const { data, error } = await supabase.rpc('generate_llms_txt_content', {
+        const { data, error } = await supabase.rpc('generate_item_llms_txt', {
           p_category: category,
           p_slug: slug,
         });
 
         if (error) {
           console.error('RPC error (item):', error);
-          return errorResponse(error, 'generate_llms_txt_content', getCorsHeaders);
+          return errorResponse(error, 'generate_item_llms_txt', getCorsHeaders);
         }
 
         if (!data) {
@@ -133,197 +133,7 @@ Deno.serve(async (req) => {
           });
         }
 
-        // Format JSONB response to plain text with full content expansion
-        const itemData = data as {
-          slug: string;
-          title: string;
-          description: string;
-          category: string;
-          tags?: string[];
-          author?: string;
-          date_added?: string;
-          url: string;
-          content?: string;
-          detailed_content?: string;
-          features?: string[];
-          use_cases?: string[];
-          examples?: Array<{ title: string; description?: string; code: string; language?: string }>;
-          metadata?: Record<string, unknown>;
-        };
-
-        const sections: string[] = [];
-        sections.push(`# ${itemData.title}\n`);
-        sections.push(`${itemData.description}\n`);
-        sections.push('---\n');
-
-        // Metadata section
-        sections.push('## Metadata\n');
-        sections.push(`**Title:** ${itemData.title}`);
-        sections.push(`**Category:** ${itemData.category}`);
-        if (itemData.author) sections.push(`**Author:** ${itemData.author}`);
-        if (itemData.date_added) {
-          const date = new Date(itemData.date_added);
-          sections.push(
-            `**Added:** ${date.toLocaleDateString('en-US', { month: 'long', year: 'numeric' })}`
-          );
-        }
-        if (itemData.tags && itemData.tags.length > 0) {
-          sections.push(`**Tags:** ${itemData.tags.join(', ')}`);
-        }
-        sections.push(`**URL:** ${SITE_URL}/${itemData.category}/${itemData.slug}\n`);
-
-        // Overview
-        sections.push('## Overview\n');
-        sections.push(`${itemData.description}\n`);
-
-        // Main content
-        if (itemData.detailed_content) {
-          sections.push('## Content\n');
-          sections.push(itemData.detailed_content);
-          sections.push('');
-        } else if (itemData.content) {
-          sections.push('## Content\n');
-          sections.push(itemData.content);
-          sections.push('');
-        }
-
-        // Features
-        if (itemData.features && itemData.features.length > 0) {
-          sections.push('## Features\n');
-          for (const feature of itemData.features) {
-            sections.push(`- ${feature}`);
-          }
-          sections.push('');
-        }
-
-        // Use Cases
-        if (itemData.use_cases && itemData.use_cases.length > 0) {
-          sections.push('## Use Cases\n');
-          for (const useCase of itemData.use_cases) {
-            sections.push(`- ${useCase}`);
-          }
-          sections.push('');
-        }
-
-        // Examples
-        if (itemData.examples && itemData.examples.length > 0) {
-          sections.push('## Examples\n');
-          for (const example of itemData.examples) {
-            sections.push(`### ${example.title}\n`);
-            if (example.description) {
-              sections.push(`${example.description}\n`);
-            }
-            sections.push('```' + (example.language || 'plaintext'));
-            sections.push(example.code);
-            sections.push('```\n');
-          }
-        }
-
-        // Metadata sections (different per category)
-        if (itemData.metadata) {
-          const meta = itemData.metadata as Record<string, unknown>;
-
-          // Installation (mcp, commands, hooks, skills, statuslines)
-          if (meta.installation) {
-            sections.push('## Installation\n');
-            const inst = meta.installation as Record<string, unknown>;
-            if (inst.claudeCode && typeof inst.claudeCode === 'object') {
-              const cc = inst.claudeCode as { steps?: string[] };
-              if (cc.steps && Array.isArray(cc.steps)) {
-                sections.push('### Claude Code\n');
-                for (const step of cc.steps) {
-                  sections.push(`${step}`);
-                }
-                sections.push('');
-              }
-            }
-            if (inst.claudeDesktop && typeof inst.claudeDesktop === 'object') {
-              const cd = inst.claudeDesktop as { steps?: string[]; configPath?: Record<string, string> };
-              if (cd.steps && Array.isArray(cd.steps)) {
-                sections.push('### Claude Desktop\n');
-                for (const step of cd.steps) {
-                  sections.push(`${step}`);
-                }
-                if (cd.configPath) {
-                  sections.push('\n**Config Paths:**');
-                  for (const [os, path] of Object.entries(cd.configPath)) {
-                    sections.push(`- ${os}: \`${path}\``);
-                  }
-                }
-                sections.push('');
-              }
-            }
-          }
-
-          // Configuration (most categories)
-          if (meta.configuration) {
-            sections.push('## Configuration\n');
-            sections.push('```json');
-            sections.push(JSON.stringify(meta.configuration, null, 2));
-            sections.push('```\n');
-          }
-
-          // Troubleshooting (most categories)
-          if (meta.troubleshooting && Array.isArray(meta.troubleshooting)) {
-            sections.push('## Troubleshooting\n');
-            for (const item of meta.troubleshooting as Array<{ issue: string; solution: string }>) {
-              sections.push(`### ${item.issue}\n`);
-              sections.push(`${item.solution}\n`);
-            }
-          }
-
-          // Collections-specific
-          if (meta.items && Array.isArray(meta.items)) {
-            sections.push(`## Collection Items (${(meta.items as unknown[]).length})\n`);
-            if (meta.installation_order && Array.isArray(meta.installation_order)) {
-              sections.push('**Installation Order:**');
-              for (let i = 0; i < (meta.installation_order as string[]).length; i++) {
-                sections.push(`${i + 1}. ${(meta.installation_order as string[])[i]}`);
-              }
-              sections.push('');
-            }
-          }
-
-          // MCP-specific
-          if (meta.permissions && Array.isArray(meta.permissions)) {
-            sections.push(`**Permissions:** ${(meta.permissions as string[]).join(', ')}\n`);
-          }
-          if (meta.requires_auth !== undefined) {
-            sections.push(`**Requires Authentication:** ${meta.requires_auth ? 'Yes' : 'No'}\n`);
-          }
-          if (meta.security && Array.isArray(meta.security)) {
-            sections.push('**Security Notes:**');
-            for (const note of meta.security as string[]) {
-              sections.push(`- ${note}`);
-            }
-            sections.push('');
-          }
-
-          // Skills-specific
-          if (meta.requirements && typeof meta.requirements === 'object') {
-            sections.push('## Requirements\n');
-            sections.push('```json');
-            sections.push(JSON.stringify(meta.requirements, null, 2));
-            sections.push('```\n');
-          }
-
-          // Guides-specific
-          if (meta.difficulty) {
-            sections.push(`**Difficulty:** ${meta.difficulty}\n`);
-          }
-          if (meta.readingTime) {
-            sections.push(`**Reading Time:** ${meta.readingTime}\n`);
-          }
-        }
-
-        sections.push('---\n');
-        sections.push(`Source: Claude Pro Directory`);
-        sections.push(`Website: ${SITE_URL}`);
-        sections.push(`URL: ${SITE_URL}/${itemData.category}/${itemData.slug}\n`);
-        sections.push('This content is optimized for Large Language Models (LLMs).');
-        sections.push('For full formatting and interactive features, visit the website.');
-
-        content = sections.join('\n');
+        content = data;
         break;
       }
 
@@ -408,14 +218,13 @@ Deno.serve(async (req) => {
       category: category || 'N/A',
       slug: slug || 'N/A',
       tool: tool || 'N/A',
-      contentLength: content.length,
+      bytes: content.length,
     });
 
     return new Response(content, {
       status: 200,
       headers: {
         'Content-Type': 'text/plain; charset=utf-8',
-        'Content-Length': content.length.toString(),
         'X-Robots-Tag': 'index, follow',
         'Cache-Control': 'public, s-maxage=3600, stale-while-revalidate=7200',
         'CDN-Cache-Control': 'max-age=3600',
